Rename custom_blocklist_checker to blocklist_checker.ts
